sessions: move aux bar toggle to layout actions, expose in command palette
Move ToggleAuxiliaryBarAction into browser/layoutActions.ts (the home of the
other sessions layout actions), drop the action's icon, and surface it in the
Command Palette under View (f1 + Categories.View). The state-dependent editor
title icon is still provided via two when-gated MenuId.EditorTitleLayout items.
Tests moved to the layout actions test suite.
