+++ src/vs/sessions/contrib/editor/test/browser/editor.contribution.test.ts
@@ -196,64 +196,4 @@ suite('Sessions - Editor Contribution', () => {
 		assert.deepStrictEqual(layoutService.maximizedStates, [true, false]);
 		assert.strictEqual(layoutService.panelVisible, true);
 	});
-
-	test('toggle auxiliary bar hides the auxiliary bar when visible', async () => {
-		const instantiationService = store.add(new TestInstantiationService());
-		const layoutService = new class extends mock<IAgentWorkbenchLayoutService>() {
-			auxiliaryBarVisible = true;
-			readonly hiddenParts: Parts[] = [];
-
-			override isVisible(part: Parts): boolean {
-				return part === Parts.AUXILIARYBAR_PART ? this.auxiliaryBarVisible : false;
-			}
-
-			override setPartHidden(hidden: boolean, part: Parts): void {
-				if (part === Parts.AUXILIARYBAR_PART) {
-					this.auxiliaryBarVisible = !hidden;
-					if (hidden) {
-						this.hiddenParts.push(part);
-					}
-				}
-			}
-		};
-		instantiationService.set(IAgentWorkbenchLayoutService, layoutService);
-
-		const handler = CommandsRegistry.getCommand('workbench.action.agentSessions.toggleAuxiliaryBar')?.handler;
-		assert.ok(handler, 'Command handler should be registered');
-
-		await handler(instantiationService);
-
-		assert.deepStrictEqual(layoutService.hiddenParts, [Parts.AUXILIARYBAR_PART]);
-		assert.strictEqual(layoutService.auxiliaryBarVisible, false);
-	});
-
-	test('toggle auxiliary bar shows the auxiliary bar when hidden', async () => {
-		const instantiationService = store.add(new TestInstantiationService());
-		const layoutService = new class extends mock<IAgentWorkbenchLayoutService>() {
-			auxiliaryBarVisible = false;
-			readonly shownParts: Parts[] = [];
-
-			override isVisible(part: Parts): boolean {
-				return part === Parts.AUXILIARYBAR_PART ? this.auxiliaryBarVisible : false;
-			}
-
-			override setPartHidden(hidden: boolean, part: Parts): void {
-				if (part === Parts.AUXILIARYBAR_PART) {
-					this.auxiliaryBarVisible = !hidden;
-					if (!hidden) {
-						this.shownParts.push(part);
-					}
-				}
-			}
-		};
-		instantiationService.set(IAgentWorkbenchLayoutService, layoutService);
-
-		const handler = CommandsRegistry.getCommand('workbench.action.agentSessions.toggleAuxiliaryBar')?.handler;
-		assert.ok(handler, 'Command handler should be registered');
-
-		await handler(instantiationService);
-
-		assert.deepStrictEqual(layoutService.shownParts, [Parts.AUXILIARYBAR_PART]);
-		assert.strictEqual(layoutService.auxiliaryBarVisible, true);
-	});
 });
